Fix linting warnings and TypeScript errors
- Fixed Fast refresh warnings by moving metadata exports to separate files
- Fixed TypeScript 'any' type warnings in monitoring.ts with proper type assertions
- Added ESLint disable comment for necessary metadata exports in layout.tsx
- All linting warnings resolved, tests passing, build successful
